==> 01_elimina_tweets_de_abril.ipynb <==
import pandas as pd
import random
import utils.functions as f
from utils.stop_words_custom import my_stop_words as stop_words
pd.set_option('display.max_colwidth', -1)

data = "data/00_initial/30.xlsx"

df = f.get_data(data)

df.info()

df.head(1)

df.tail(1)

df['pubdate'] = pd.to_datetime(df['pubdate'])

df.info()

df.set_index('pubdate', inplace=True)
df.info()

df.head()

no_april = df.loc['2019-05-01':'2019-05-31']
no_april.head()

df.shape

no_april.shape

len(df) - len(no_april)

writer = pd.ExcelWriter(r'data/01_no_april/30.xlsx', engine='xlsxwriter',options={'strings_to_urls': False})
no_april.to_excel(writer)
writer.close()
