==> 03_concatena_arquivos_dos_dias.ipynb <==
import pandas as pd
import random
import utils.functions as f
from utils.stop_words_custom import my_stop_words as stop_words
pd.set_option('display.max_colwidth', -1)

df30 = f.get_data("data/01_no_april/30.xlsx")
df31 = f.get_data("data/00_initial/31.xlsx")
df32 = f.get_data("data/00_initial/32.xlsx")
df33 = f.get_data("data/00_initial/33.xlsx")
df34 = f.get_data("data/00_initial/34.xlsx")
df35 = f.get_data("data/00_initial/35.xlsx")
df36 = f.get_data("data/00_initial/36.xlsx")
df37 = f.get_data("data/00_initial/37.xlsx")
df38 = f.get_data("data/00_initial/38.xlsx")
df39 = f.get_data("data/00_initial/39.xlsx")
df40 = f.get_data("data/02_no_june/40.xlsx")

df30.info()

df31.info()

df32.info()

df33.info()

df34.info()

df35.info()

df36.info()

df37.info()

df38.info()

df39.info()

df40.info()

df_all = pd.concat([df30,df31,df32,df33,df34,df35,df36,df37,df38,df39,df40], sort=False)

df_all.head()

df_all.info()

df_all.shape

writer = pd.ExcelWriter(r'data/03_just_may/may.xlsx', engine='xlsxwriter',options={'strings_to_urls': False})
df_all.to_excel(writer)
writer.close()
