==> 04_EDA.ipynb <==
import pandas as pd
import random
import re
import utils.functions as f
from utils.stop_words_custom import my_stop_words as stop_words
pd.set_option('display.max_colwidth', -1)

df = f.get_data("data/03_just_may/may.xlsx")

df.columns

locations = df['user_location'].value_counts(dropna=False)
locations.iloc[0:50]

index = 0
new_pubdate = []
for row in df['user_location']:
    if type(row) == str:
        if re.findall(r'[Ss][ãa]o [Pp]aulo.*', row):
            pubdate = re.sub(r'[Ss][ãa]o [Pp]aulo.*', "São Paulo", row)
        elif re.findall(r'^[Ss].*[Pp]aulo.*', row):
            pubdate = re.sub(r'^[Ss].*[Pp]aulo.*', "São Paulo", row)
        elif re.findall(r'^[Rr]io de [Jj]aneiro.*', row):
            pubdate = re.sub(r'^[Rr]io de [Jj]aneiro.*', "Rio de Janeiro", row)
#         elif re.findall(r'.*[Rr][Jj].*', row):
#             pubdate = re.sub(r'.*[Rr][Jj].*', "Rio de Janeiro", row)
        elif re.findall(r'^[Bb]ra[sz]il.*', row):
            pubdate = re.sub(r'^[Bb]ra[sz]il.*', "Brasil", row)
        elif re.findall(r'^[Bb]elo [Hh]orizonte.*', row):
            pubdate = re.sub(r'^[Bb]elo [Hh]orizonte.*', "Belo Horizonte", row)
        elif re.findall(r'^[Bb]ras[ií]lia.*', row):
            pubdate = re.sub(r'^Bras[ií]lia.*', "Brasília", row)
        elif re.findall(r'^[Cc]uritiba.*', row):
            pubdate = re.sub(r'[Cc]uritiba.*', "Curitiba", row)
        elif re.findall(r', [Bb]ra[zs]il.*', row):
            pubdate = re.sub(r', [Bb]ra[zs]il.*', "", row)
        else:
            pubdate = row
    else:
        pubdate = row
    new_pubdate.append(pubdate)
    index = index + 1

pd.Series(new_pubdate).value_counts(dropna=False).iloc[0:10]

# df['title'].value_counts().iloc[0:10]

len(df['title'].unique())

# df['description'].value_counts().iloc[0:20]

len(df['description'].unique())

df['author'].value_counts(dropna=False)

df['user_created_at'].value_counts(dropna=False)

df['user_friends_count'].mean()

df['user_friends_count'].median()

df['user_followers_count'].mean()

df['user_followers_count'].median()

criteria = df['user_created_at'] == '2018-12-18 18:11:05'
df[criteria]["description"].iloc[0:10]

df['user_bio'].value_counts().iloc[0:50]

criteria = df['user_bio'] == "#lulalivre"
teste = df[criteria]["description"]
teste.iloc[0:10]

criteria = df['user_bio'] == "#LulaLivre"
teste = df[criteria]["description"]
teste.iloc[0:10]

criteria = df['user_bio'] == "«Todas las pistas que llevan a Bolsonaro» #LulaLivre #LulaInocente #QuemMatouMarielle #BrésilSousRégimeFasciste"
teste = df[criteria]["description"]
teste.iloc[0:10]

df['source'].value_counts().iloc[0:50]

#df.set_index('pubdate', inplace=True)
