==> 06_elimina_rts.ipynb <==
import pandas as pd
import random
import re
import utils.functions as f
from utils.stop_words_custom import my_stop_words as stop_words
pd.set_option('display.max_colwidth', -1)

df = f.get_data('data/05_ordered_may/may_ordered.xlsx')

no_rt = []
for index, row in df.iterrows():
    if not(str(row["title"]).startswith("RT")):
        no_rt.append(row)    

df.shape

len(no_rt)

df.shape[0] - len(no_rt)

df_no_rt = pd.DataFrame(no_rt)

writer = pd.ExcelWriter(r'data/06_rts_deleted/without_rts.xlsx', engine='xlsxwriter',options={'strings_to_urls': False})
df_no_rt.to_excel(writer)
writer.close()
