==> 07_elimina_descriptions_repetidas.ipynb <==
import pandas as pd
import random
import re
import utils.functions as f
from utils.stop_words_custom import my_stop_words as stop_words
pd.set_option('display.max_colwidth', -1)

df = f.get_data('data/06_rts_deleted/without_rts.xlsx')

no_repeated_description = []
list_descriptions = []
for index, row in df.iterrows():
    if row["description"] not in list_descriptions :
        no_repeated_description.append(row)    
    list_descriptions.append(row["description"])

df.shape

len(no_repeated_description)

df.shape[0] - len(no_repeated_description)

df_no_repeated_description = pd.DataFrame(no_repeated_description)

writer = pd.ExcelWriter(r'data/07_repeated_descriptions_deleted/without_repeated_descriptions.xlsx', engine='xlsxwriter',options={'strings_to_urls': False})
df_no_repeated_description.to_excel(writer)
writer.close()
