==> 08_separacao_de_dias_em_arquivos.ipynb <==
import pandas as pd
import random
import re
import utils.functions as f
from utils.stop_words_custom import my_stop_words as stop_words
pd.set_option('display.max_colwidth', -1)

df = f.get_data('data/07_repeated_descriptions_deleted/without_repeated_descriptions.xlsx')

df.set_index('pubdate', inplace=True)

df.info()

df.head()

dict_bolsonaro = {}
for i in range(1,32):  
    if(i<10):      
        start = '2019-05-0{}'.format(str(i))
        end   = '2019-05-0{}'.format(str(i))
        name = '2019-05-0{}'.format(str(i))
        dict_bolsonaro[name] = df[start:end]
    else:
        start = '2019-05-{}'.format(str(i))
        end   = '2019-05-{}'.format(str(i))
        name = '2019-05-{}'.format(str(i))
        dict_bolsonaro[name] = df[start:end]
        
    writer = pd.ExcelWriter(r'data/08_separated_day_files/{}.xlsx'.format(name), engine='xlsxwriter',options={'strings_to_urls': False})
    dict_bolsonaro[name].to_excel(writer)
    writer.close()
