==> 09_columns_deleteds.ipynb <==
import pandas as pd
import random
import re
import utils.functions as f
from utils.stop_words_custom import my_stop_words as stop_words
pd.set_option('display.max_colwidth', -1)

df = f.get_data('data/07_repeated_descriptions_deleted/without_repeated_descriptions.xlsx')

df.info()

df.head(1)

df.drop(['guid','link','title','user_id','profile_image_url','lang','code','coords','source','user_statuses_count','user_friends_count','user_followers_count','user_location'], axis=1, inplace=True)
# source	user_statuses_count	user_friends_count	user_followers_count	user_created_at	user_bio	user_location

df.info()

df.head()

df["author"].value_counts().head(10)

df["user_bio"].value_counts().head(50)

df["user_created_at"].value_counts().head(10)

writer = pd.ExcelWriter(r'data/09_columns_deleteds/columns_deleteds.xlsx', engine='xlsxwriter',options={'strings_to_urls': False})
df.to_excel(writer)
writer.close()
