==> 10_inicio_classificacao_tweets_treinamento.ipynb <==
import pandas as pd
import random
import re
import utils.functions as f
from utils.stop_words_custom import my_stop_words as stop_words
pd.set_option('display.max_colwidth', -1)

df = f.get_data('data/09_columns_deleteds/columns_deleteds.xlsx')

df.info()

df.dropna(subset=['description'],inplace=True)

classified  = []
description = []
classes     = []
positive    = []
negative    = []
neutral     = []
i           = 0

users_bio_negative = ["Manifestem-se! Só um velho engenheiro econômico aposentado, um vagabundo a mais, segundo um conhecido sociopata canalha! #EsquerdistasSeguemEsquerdistas","Engenheiro Civil. Petista,Lulista,Pacifista,Anti-Preconceitos,Corintiano e sobretudo BRASILEIRO.#LULALIVRE","Precisamos vencer a fome, a miséria e a exclusão social. Nossa guerra não é para matar ninguém - é para salvar vidas.\n\nLuiz Inácio Lula da Silva","CASADO, PAI DE 3, AVÔ DE 2 ATÉ AGORA, CATÓLICO, FLAMENGUISTA, ENGo.TELECOM, BRIZOLISTA RADICALMENTE ANTI-DIREITA!   #EsquerdistasSeguemEsquerdistas"                     
    ]

for index,data in df.iterrows():
    row = data["description"]
    if isinstance(data["user_bio"],str):
        if("#Dia15GreveGeralNaEducação" in row):
            negative.append(row)
            description.append(row)
            classes.append("negative")
        elif(row in users_bio_negative):
            negative.append(row)
            description.append(row)
            classes.append("negative")
        elif("#ArmasMatam" in row):
            negative.append(row)
            description.append(row)
            classes.append("negative")
        elif("#NaoVaiTerCorteVaiTerLuta" in row):
            negative.append(row)
            description.append(row)
            classes.append("negative")
        elif("#EleNao" in row):
            negative.append(row)
            description.append(row)
            classes.append("negative")
        elif("#ForaBolsonaro" in row):
            negative.append(row)
            description.append(row)
            classes.append("negative")
        elif("#CadêOQueiroz" in row):
            negative.append(row)
            description.append(row)
            classes.append("negative")
        elif("#SomosTodosLeticia" in row):
            negative.append(row)
            description.append(row)
            classes.append("negative")
        elif("#CadeOQueiroz" in row):
            negative.append(row)
            description.append(row)
            classes.append("negative")
        elif("#BolsonaroEnvergonhaOBrasil" in row):
            negative.append(row)
            description.append(row)
            classes.append("negative")
        elif("#UniversidadesContraBolsonaro" in row):
            negative.append(row)
            description.append(row)
            classes.append("negative")
        elif("#CancelBolsonaro" in row):
            negative.append(row)
            description.append(row)
            classes.append("negative")
        elif("#PersonulidadeDoAno" in row):
            negative.append(row)
            description.append(row)
            classes.append("negative")
        elif("#LulaLivre" in row):
            negative.append(row)
            description.append(row)
            classes.append("negative")
        elif("#LulaInspira" in row):
            negative.append(row)
            description.append(row)
            classes.append("negative")
        elif("#takeitdownMarriott" in row):
            negative.append(row)
            description.append(row)
            classes.append("negative")
        elif("#Bolsonaro2022" in row):
            positive.append(row)
            description.append(row)
            classes.append("positive")
        elif("#BrasilÉBolsonaro" in row):
            positive.append(row)
            description.append(row)
            classes.append("positive")
        elif("#PTNuncaMais" in row):
            positive.append(row)
            description.append(row)
            classes.append("positive")
        elif("#elesim" in row):
            positive.append(row)
            description.append(row)
            classes.append("positive")
        elif("#DireitaUnida" in row):
            positive.append(row)
            description.append(row)
            classes.append("positive")
        elif("#EstouComBolsonaro" in row):
            positive.append(row)
            description.append(row)
            classes.append("positive")
        elif("#BolsonaroOrgulhodoBrasil" in row):
            positive.append(row)
            description.append(row)
            classes.append("positive")
        elif("#BolsoSilvioVemAi" in row):
            positive.append(row)
            description.append(row)
            classes.append("positive")
        elif("#BolsoNoSBT" in row):
            positive.append(row)
            description.append(row)
            classes.append("positive")
        elif("#BoicoteBurgerKing" in row):
            positive.append(row)
            description.append(row)
            classes.append("positive")
        elif("#CentraoBlocoDeLadrao" in row):
            positive.append(row)
            description.append(row)
            classes.append("positive")
        elif("#B2022" in row):
            positive.append(row)
            description.append(row)
            classes.append("positive")
    i = i+ 1

classified = pd.DataFrame(classified)
classified["description"] = description
classified["class"] = classes

classified.head()

classified["class"].value_counts()

for tweet in positive[0:50]:
    print(tweet)
    print("-------------")

writer = pd.ExcelWriter(r'data/10_classify_tweets/classify_tweets.xlsx', engine='xlsxwriter',options={'strings_to_urls': False})
classified.to_excel(writer)
writer.close()
