==> 11_finalizacao_de_classificacao_automatica_de_tweets.ipynb <==
import pandas as pd
import random
import re
import utils.functions as f
from utils.stop_words_custom import my_stop_words as stop_words
pd.set_option('display.max_colwidth', -1)

df_all = f.get_data('data/09_columns_deleteds/columns_deleteds.xlsx')

df_all.dropna(subset=['description'],inplace=True)

df_all.info()

# Desfavoráveis
marielle = []
queiroz  = []
milicia  = []
balburdia  = []
reeleicao = []
minion = []
tomar_no_cu = []
golpe = []
ditadura = []
ustra = []
lula_livre = []
preso_politico = []
corte_de_verbas = []
bozo = []
fascis = []
nazis = []
idiotas_uteis = []
moro_vendido = []
vaga = []
laranja = []
educacao = []
greve = []
contra_a_reforma = []
vai_ter_luta = []
cuba = []

# Favoráveis
mortadela = []
bolsonaro_warrior = []
porte_armas = []
brasil_weintraub =[]
bolsonaro_nosso_presidente = []
bolsonaro_tem_razao = []
deus_brasil_bolsonaro = []
dia_26_eu_vou = []
dia_26_brasil_ruas = []
nordeste_com_bolsonaro =[]
domingo_pelo_brasil =[]
quem_mandou_matar_bolsonaro =[]
bolsonaro_no_the_noite =[]


bom_presidente = []
bolsonaro_ser = []

for index,row in df_all.iterrows():
    if "marielle" in str(row["description"]).lower():
        marielle.append(row["description"])
    elif "queiroz" in str(row["description"]).lower():
        queiroz.append(row["description"])
    elif "milícia" in str(row["description"]).lower():
        milicia.append(row["description"])
    elif "balbúrdia" in str(row["description"]).lower():
        balburdia.append(row["description"])
    elif "2022" in str(row["description"]).lower():
        reeleicao.append(row["description"])
    elif "minion" in str(row["description"]).lower():
        minion.append(row["description"])
    elif "tomar no cú" in str(row["description"]).lower() or "tomar no cu" in str(row["description"]).lower():
        tomar_no_cu.append(row["description"])
    elif "golpe" in str(row["description"]).lower():
        golpe.append(row["description"])
    elif "ditadura" in str(row["description"]).lower():
        ditadura.append(row["description"])
    elif "ustra" in str(row["description"]).lower():
        ditadura.append(row["description"])
    elif "lulalivre" in str(row["description"]).lower():
        lula_livre.append(row["description"])
    elif "preso político" in str(row["description"]).lower():
        preso_politico.append(row["description"])
    elif "corte de verbas" in str(row["description"]).lower():
        corte_de_verbas.append(row["description"])
    elif "bozo" in str(row["description"]).lower():
        bozo.append(row["description"])
    elif "fascis" in str(row["description"]).lower():
        fascis.append(row["description"])
    elif "nazis" in str(row["description"]).lower():
        nazis.append(row["description"])
    elif "idiotas úteis" in str(row["description"]).lower():
        idiotas_uteis.append(row["description"])
    elif "#morovendido" in str(row["description"]).lower():
        moro_vendido.append(row["description"])
    elif "vaga " in str(row["description"]).lower():
        vaga.append(row["description"])
    elif "laranja" in str(row["description"]).lower():
        laranja.append(row["description"])
    elif "#tsunaminaeducação" in str(row["description"]).lower():
        educacao.append(row["description"])
    elif "greve" in str(row["description"]).lower():
        greve.append(row["description"])
    elif "contra a reforma da previdência" in str(row["description"]).lower():
        contra_a_reforma.append(row["description"])
    elif "vai ter luta" in str(row["description"]).lower():
        vai_ter_luta.append(row["description"])
    elif "cuba" in str(row["description"]).lower():
        cuba.append(row["description"])
    ###############################################################################################################
    elif "bolsonaro é" in str(row["description"]).lower():
        bolsonaro_ser.append(row["description"])
    ##############################################################################################################
    elif "mortadela" in str(row["description"]).lower():
        mortadela.append(row["description"])
    elif "#bolsonarowarrior" in str(row["description"]).lower():
        bolsonaro_warrior.append(row["description"])
    elif "#brasilexigeportedearmas" in str(row["description"]).lower():
        porte_armas.append(row["description"])
    elif "#brasilcomweintraub" in str(row["description"]).lower():
        brasil_weintraub.append(row["description"])
    elif "#bolsonaronossopresidente" in str(row["description"]).lower():
        bolsonaro_nosso_presidente.append(row["description"])
    elif "#BolsonaroTemRazao".lower() in str(row["description"]).lower():
        bolsonaro_tem_razao.append(row["description"])
    elif "#DeusBrasilBolsonaro".lower() in str(row["description"]).lower():
        deus_brasil_bolsonaro.append(row["description"])
    elif "#Dia26EuVou".lower() in str(row["description"]).lower():
        dia_26_eu_vou.append(row["description"])
    elif "#Dia26BrasilNasRuas".lower() in str(row["description"]).lower():
        dia_26_brasil_ruas.append(row["description"])
    elif "#NordesteComBolsonaro".lower() in str(row["description"]).lower():
        nordeste_com_bolsonaro.append(row["description"])
    elif "#DomingoPeloBrasil".lower() in str(row["description"]).lower():
        domingo_pelo_brasil.append(row["description"])
    elif "#QuemMandouMatarBolsonaro".lower() in str(row["description"]).lower():
        quem_mandou_matar_bolsonaro.append(row["description"])
    elif "#BolsonaroNoTheNoite".lower() in str(row["description"]).lower():
        bolsonaro_no_the_noite.append(row["description"])
    elif "bom presidente" in str(row["description"]).lower():
        bom_presidente.append(row["description"])

len(mortadela)

bolsonaro_warrior[0:10]

len(bolsonaro_nosso_presidente)

negative = marielle + queiroz + milicia + balburdia + minion + tomar_no_cu + golpe + ditadura + lula_livre
negative = negative + preso_politico + corte_de_verbas + bozo + fascis + nazis + idiotas_uteis + mortadela 
negative = negative + moro_vendido + vaga + laranja + educacao + greve + contra_a_reforma + vai_ter_luta

len(negative)

positive = mortadela + bolsonaro_warrior + porte_armas + brasil_weintraub + bolsonaro_nosso_presidente + reeleicao
positive = positive + bolsonaro_tem_razao + deus_brasil_bolsonaro + dia_26_eu_vou + dia_26_brasil_ruas + nordeste_com_bolsonaro
positive = positive + domingo_pelo_brasil + quem_mandou_matar_bolsonaro + bolsonaro_no_the_noite + cuba

len(positive)

def unique(list1): 
    unique_list = [] 
    for x in list1: 
        if x not in unique_list: 
            unique_list.append(x) 
    return unique_list

uniques = unique(negative)
len(uniques)

len(marielle)

len(queiroz)

len(milicia)

len(balburdia)

len(reeleicao)

len(minion)

len(tomar_no_cu)

len(golpe)

len(ditadura)

len(lula_livre)

len(preso_politico)

len(corte_de_verbas)

len(bozo)

len(fascis)

len(nazis)

len(idiotas_uteis)

len(moro_vendido)

len(vaga)

len(laranja)

len(educacao)

len(mortadela)

len(bom_presidente)

len(greve)

len(contra_a_reforma)

len(vai_ter_luta)

len(cuba)

len(bolsonaro_ser)

len(bolsonaro_warrior)

bolsonaro_ser[0:50]

nazis[0:10]

fascis[0:10]

len(negative)

len(positive)

writer = pd.ExcelWriter(r'data/11_classified/negative.xlsx', engine='xlsxwriter',options={'strings_to_urls': False})
pd.DataFrame(negative).to_excel(writer)
writer.close()

writer = pd.ExcelWriter(r'data/11_classified/positive.xlsx', engine='xlsxwriter',options={'strings_to_urls': False})
pd.DataFrame(positive).to_excel(writer)
writer.close()
